Fix transceiver data corruption
Because of different source and destination dtypes in numpy,
the receiver corrupted the data and interpreted it in a wrong
way. In addition, the background provider was moved to the
external process in order to save memory (no need to have a
copy of the background in the main process, before the fork).